Fixed a bug in Connect-DSCC command
Bug found by Peter R (thanks), where the connect command would throw an error if the DSCC instance did not have any host-initiator-groups defined. The connect command uses a single command (in this case host-initiator-groups) to determine if it can communicate with the region the token was gathered for. This command returned NULL, and the connect command was determining that was a failed retrieval instead of a successfull command with a count of zero.