Fixed StorageSystem and Initiators
The Get-DSCCStorageSystem command added an option to define by name, and to set a later used variable, however this added code started throwing errors in the test enviornemnt. Additionally the API has changed around the process of Creating Initiators, and New-DSCCInitiator was fixed to disallow creating FC initiators for DeviceType1 since those are now only autodetected and not created by the API. Also fixed the example helps which had incorrect command names.