Updated to require PowerShell 7.
There were a mounting number of bugs and code branches that can be eliminated by simply using an up to date version of PowerShell. Version 5.x of powershell offered too many dead ends and lack of features. i.e. The test-connection was lacking. New codebase will require 7, and refuse to load with an appropriate message if using powershell < 7